+++ packages/replicate/src/index.ts
@@ -3,4 +3,5 @@ export type {
   ReplicateProvider,
   ReplicateProviderSettings,
 } from './replicate-provider';
+export type { ReplicateImageProviderOptions } from './replicate-image-model';
 export { VERSION } from './version';