+++ packages/test-server/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ai-sdk/test-server
 
+## 1.0.0
+
+### Major Changes
+
+- dee8b05: ai SDK 6 beta
+- 953d0f2: feat(packages/test-server): Add `test-server` as a package
+
 ## 1.0.0-beta.1
 
 ### Major Changes